tests/requirements.txt
tox
pytest-metadata